pyproject.toml
[tool.poetry]
name = "serverless-builder"
version = "0.1.0"
description = ""
authors = ["Michal Przytulski <michal@przytulski.pl>"]

packages = [
    { include = "serverless" }
]

[tool.poetry.dependencies]
python = "^3.6.2"
PyYAML = "^6.0"
troposphere = "^3.1.1"
stringcase = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.12b0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
